#6 - permutations; #3 - fixing waiting model
- fix for edges model : When switching directions at the end of the
line, I have to wait. Before this fix, an edge used to connect second
last station directly to the end of line  node with train arrived. This
bypassed the usual flow, in which one, in average, has to wait some time
for the train back.
- new function to allow for checkpoints to be defined for paths. A path
is constructed from a sequence of shortest paths which connect the given
checkpoints in the given order. A total time is provided, along with
stations not covered by the path
- since joining line ends with shortest paths might not cover all
stations - our original problem - a function was added to check for
missing stations on the proposed path.
- only invoking nuget if quickgraph is missing